[Add] 캐릭터 SD, 타일 스프라이트 추가
Layer sorting 설정 변경

+++ Scissors_Tale/Assets/Scripts/Utils/Utils.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public static class Utils
 {
     //참고용
-    public const int TileSize = 1;  // 타일의 실제 크기
+    public const float TileSize = 1f;  // 타일의 실제 크기
     // 체스판의 크기
     public const int FieldWidth = 5;
     public const int FieldHeight = 5;
